+++ tests/test.ts
@@ -1,6 +1,0 @@
-import { Dexonline } from '../mod.ts';
-
-Deno.test('parser', async () => {
-	const entries = await Dexonline.get('încă');
-	console.debug(entries);
-});