+++ CHANGELOG.md
@@ -1,1 +1,18 @@
 ## 1.0.0 (Unreleased)
+
+FEATURES:
+
+- Options:
+  - Modes:
+    - Strict (Match identical)
+    - Lax (Match similar)
+- Parsing entries in the 'Synthesis' tab:
+  - Headers:
+    - Lemma
+    - Type
+  - Body:
+    - Examples
+    - Definitions
+    - Expressions
+    - Relations
+    - Etymology